# Conflicts:
#	app/containers/Inventory/actions.js
#	app/containers/Inventory/constants.js
#	app/containers/Inventory/reducer.js
#	app/containers/Inventory/saga.js
#	app/containers/NavBar/index.js
#	app/containers/RestaurantList/index.js
#	app/containers/ShoppingCart/actions.js
#	app/containers/ShoppingCart/index.js
#	app/containers/ShoppingCart/reducer.js
#	app/containers/ShoppingCart/saga.js
#	database/migrations/01_recipes.js
#	server/inventory.js

+++ database/migrations/01_recipes.js
@@ -14,7 +14,7 @@ exports.up = knex =>
       table.string('restaurant_phone_number').notNullable();
       table.string('restaurant_website');
       table.timestamp('created_at').defaultTo(knex.fn.now());
-      table.integer('user_id'); // foreign('user_id').references('users.id');
+      table.bigInteger('user_id'); // foreign('user_id').references('users.id');
     })
 
     // 02 RECIPES
